refactor: dialog 패키지 폴더 변경
